support for building on non-linux operating systems
