+++ config.go
@@ -9,26 +9,13 @@ package zeroconfig
 import (
 	"fmt"
 	"io"
-	"log/syslog"
 	"os"
 	"sort"
 	"time"
 
 	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/journald"
 	"gopkg.in/natefinch/lumberjack.v2"
 )
-
-// SyslogConfig contains the configuration options for the syslog writer.
-//
-// See https://pkg.go.dev/log/syslog for exact details.
-type SyslogConfig struct {
-	// All parameters are passed to https://pkg.go.dev/log/syslog#Dial directly.
-	Network string          `json:"network,omitempty" yaml:"network,omitempty"`
-	Host    string          `json:"host,omitempty" yaml:"host,omitempty"`
-	Flags   syslog.Priority `json:"flags,omitempty" yaml:"flags,omitempty"`
-	Tag     string          `json:"tag,omitempty" yaml:"tag,omitempty"`
-}
 
 // FileConfig contains the configuration options for the file writer.
 //
@@ -59,13 +46,6 @@ const (
 	// WriterTypeFile writes to a file, including rotating the file when it gets big.
 	// The configuration is stored in the FileConfig struct.
 	WriterTypeFile WriterType = "file"
-	// WriterTypeSyslog writes to the system logging service.
-	// The configuration is stored in the SyslogConfig struct.
-	WriterTypeSyslog WriterType = "syslog"
-	// WriterTypeSyslogCEE writes to the system logging service with MITRE CEE prefixes.
-	WriterTypeSyslogCEE WriterType = "syslog-cee"
-	// WriterTypeJournald writes to systemd's logging service.
-	WriterTypeJournald WriterType = "journald"
 )
 
 // LogFormat describes how logs should be formatted for a writer.
@@ -133,20 +113,8 @@ func (wc *WriterConfig) compileMain() (io.Writer, error) {
 			return nil, err
 		}
 		return writer, nil
-	case WriterTypeSyslog, WriterTypeSyslogCEE:
-		sl, err := syslog.Dial(wc.Network, wc.Host, wc.Flags, wc.Tag)
-		if err != nil {
-			return nil, err
-		}
-		if wc.Type == WriterTypeSyslogCEE {
-			return zerolog.SyslogCEEWriter(sl), nil
-		} else {
-			return zerolog.SyslogLevelWriter(sl), nil
-		}
-	case WriterTypeJournald:
-		return journald.NewJournalDWriter(), nil
 	default:
-		return nil, fmt.Errorf("unknown writer type %q", wc.Type)
+		return wc.compileMainForOS()
 	}
 }
